requirements.txt
streamlit
fpdf
python-dotenv
requests
langchain
accelerate
langchain-community
optimum
huggingface_hub
